move logic from normalize to language variants decorator
