+++ src/decorators/languageVariantsDecorator.js
@@ -1,4 +1,6 @@
-const normalize = require(`../normalize`);
+const _ = require('lodash');
+
+const validation = require('../validation');
 
 /**
  * Add Gatsby relations from item nodes to it other language variants.
@@ -15,7 +17,7 @@ const decorateItemsWithLanguageVariants = (
     of nonDefaultLanguageItemNodes) {
     defaultCultureContentItemNodes.forEach((contentItemNode) => {
       try {
-        normalize.decorateItemNodeWithLanguageVariantLink(
+        decorateItemNodeWithLanguageVariantLink(
           contentItemNode,
           currentLanguageNodes);
       } catch (error) {
@@ -28,11 +30,11 @@ const decorateItemsWithLanguageVariants = (
       if (otherLanguageCodename !== languageCodename) {
         currentLanguageNodes.forEach((contentItemNode) => {
           try {
-            normalize.decorateItemNodeWithLanguageVariantLink(
+            decorateItemNodeWithLanguageVariantLink(
               contentItemNode,
               otherLanguageNodes);
 
-            normalize.decorateItemNodeWithLanguageVariantLink(
+            decorateItemNodeWithLanguageVariantLink(
               contentItemNode,
               defaultCultureContentItemNodes);
           } catch (error) {
@@ -43,6 +45,36 @@ const decorateItemsWithLanguageVariants = (
     }
   }
 };
+
+/**
+ * Adds links between a Gatsby content item node and its
+ *    language variant nodes (translations).
+ * @param {object} itemNode - Gatsby content item node.
+ * @param {array} allNodesOfAnotherLanguage - The whole set of Gatsby item nodes
+ *    of another language.
+ * @throws {Error}
+ */
+const decorateItemNodeWithLanguageVariantLink =
+  (itemNode, allNodesOfAnotherLanguage) => {
+    validation.checkItemsObjectStructure([itemNode]);
+    validation.checkItemsObjectStructure(allNodesOfAnotherLanguage);
+
+    const languageVariantNode = allNodesOfAnotherLanguage.find(
+      (nodeOfSpecificLanguage) =>
+        itemNode.system.codename === nodeOfSpecificLanguage.system.codename
+        && itemNode.system.type === nodeOfSpecificLanguage.system.type
+        && itemNode.system.language !== nodeOfSpecificLanguage.system.language
+    );
+
+    const otherLanguageLink = languageVariantNode &&
+      itemNode.otherLanguages___NODE.find(
+        (otherLanguageId) => otherLanguageId === languageVariantNode.id
+      );
+
+    if (!otherLanguageLink && _.get(languageVariantNode, 'id')) {
+      itemNode.otherLanguages___NODE.push(languageVariantNode.id);
+    }
+  };
 
 module.exports = {
   decorateItemsWithLanguageVariants,
